added simple factory method for parsing the templates
